server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.1",
    "argon2-cffi>=25.1.0",
    "email-validator>=2.3.0",
    "facenet-pytorch>=2.5.3",
    "fastapi>=0.128.0",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "passlib[email]>=1.7.4",
    "pillow>=12.1.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "sqlalchemy>=2.0.46",
    "torch>=2.10.0",
    "torchvision>=0.25.0",
    "uvicorn[standard]>=0.40.0",
]
